requirements.txt
-e git+https://github.com/yoophi/flask-api-app.git@80a1f354574b7dbc7fdb410d0fdb49192e2f3fd1#egg=Flask_Api_App
-e git+https://github.com/yoophi/flask-config-helper.git@8c913d60226989efea68442855b301f4dc7d7822#egg=Flask_Config_Helper-master
-e git+https://github.com/yoophi/insta_client.git#egg=insta_client
-e git+https://github.com/yoophi/ipa.git#egg=ipa
-e git+https://github.com/yoophi/flask-swagger-ui.git#egg=flask_swagger_ui
Flask-Admin==1.4.2
Flask-Bootstrap==3.3.6.0
Flask-DebugToolbar==0.10.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-OAuth==0.12
Flask-OAuthlib==0.9.3
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.1
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-WTF==0.13.1
Flask==0.11.1
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==1.0.15
WTForms==2.1
Werkzeug==0.11.11
alembic==0.8.7
blinker==1.4
celery
click==6.6
dominate==2.2.1
flask-marshmallow==0.7.0
flask-uploads
flask_admin_s3_upload
flask_thumbnails_s3
future==0.15.2
httplib2==0.9.2
itsdangerous==0.24
marshmallow==2.10.0
marshmallow-sqlalchemy==0.12.1
oauth2==1.9.0.post1
oauthlib==2.0.0
passlib==1.6.5
psycopg2==2.6.2
python-editor==1.0.1
python-twitter==3.1
requests-oauthlib==0.6.2
requests==2.11.1
s3_saver
simplejson==3.8.2
six==1.10.0
visitor==0.1.3
wheel==0.24.0
bcrypt==3.1.1
flask-reggie==0.0.2
Flask-Cors==3.0.2
redis==2.10.5
